Merge branch 'master' of https://bitbucket.org/rcketscientist/rawdroid

Conflicts:
	AndroidManifest.xml
	res/xml/changelog.xml
	src/com/anthonymandra/rawdroid/CameraImport.java

+++ src/com/anthonymandra/framework/GalleryActivity.java
@@ -275,10 +275,13 @@ public abstract class GalleryActivity extends SherlockFragmentActivity
 			public void run()
 			{
 				final File[] swapFiles = mSwapDir.listFiles();
-				for (File toDelete : swapFiles)
-				{
-					toDelete.delete();
-				}
+                if (swapFiles != null)
+                {
+                    for (File toDelete : swapFiles)
+                    {
+                        toDelete.delete();
+                    }
+                }
 			}
 		}.run();
 	}
